add a check of impossible flips

+++ src/search/bfs.rs
@@ -62,7 +62,7 @@ fn process_board(
         return;
     }
     while b != 0 {
-        let index = b.trailing_zeros(); // 0..=63
+        let index = b.trailing_zeros() as usize; // 0..=63
         b &= b - 1;
 
         // “直前に相手が index に置いた” と想定したときの可能 flip 集合を列挙
